+++ core/Environment/ExecutionEnvironment.php
@@ -18,10 +18,10 @@
  * @link      https://github.com/dankocherga/MTool
  */
 
-namespace Core;
+namespace Core\Environment;
 
 /**
- * Environment interface
+ * Script execution environment
  *
  * @category Core
  * @package  Environment
@@ -29,12 +29,15 @@ namespace Core;
  * @license  http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link     https://github.com/dankocherga/MTool
  */
-interface IEnvironment
+class ExecutionEnvironment implements IEnvironment
 {
     /**
      * Get working dir 
      * 
      * @return string
      */
-    public function getWorkingDir();
+    public function getWorkingDir()
+    {
+        return getcwd();
+    }
 }
