+++ tests/core/Environment/ExecutionEnvironmentTest.php
@@ -10,34 +10,36 @@
  * that is available through the world-wide-web at this URL:
  * http://opensource.org/licenses/osl-3.0.php
  *
- * @category  Core
- * @package   Environment
+ * @category  Tests
+ * @package   Core
  * @author    Daniel Kocherga <dan.kocherga@gmail.com>
  * @copyright 2013 Daniel Kocherga (dan.kocherga@gmail.com)
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      https://github.com/dankocherga/MTool
  */
 
-namespace Core;
+namespace Core\Environment;
 
 /**
- * Script execution environment
+ * Execution environment test case
  *
- * @category Core
- * @package  Environment
+ * @category Tests
+ * @package  Core
  * @author   Daniel Kocherga <dan@oggettoweb.com>
  * @license  http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link     https://github.com/dankocherga/MTool
  */
-class ExecutionEnvironment implements IEnvironment
+class ExecutionEnvironmentTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * Get working dir 
+     * Returns current dir as working dir 
      * 
-     * @return string
+     * @return void
+     * @test
      */
-    public function getWorkingDir()
+    public function returnsCurrentDirAsWorkingDir()
     {
-        return getcwd();
+        $env = new ExecutionEnvironment;
+        $this->assertEquals(getcwd(), $env->getWorkingDir());
     }
 }
